Replaces install method of gnome-extensions-cli from yay to pip
Previously the gnome-extensions-cli was installed from the AUR with yay, but since we've migrated Python and Pip to mise-en-place, this installation method no longer works. The package published on the AUR is incompatible with the mise-en-place pip shims. To fix this, we directly install the extension with pip instead of using the AUR package.